storage/scripts/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
torchvision==0.13.1+cu113
networkx==3.1
matplotlib
scikit-image
pytorch-lightning==1.9.3
diffusers==0.31.0
einops
Pillow
webdavclient3